Moved PlanRequest to its own header file.
